update & fix tests 8)
